bug fix - publisher receiving it's own message code refactoring
