+++ src/main/java/com/example/messagingrabbitmq/configs/LocalReceiverConfig.java
@@ -96,7 +96,7 @@ public class LocalReceiverConfig {
 
   @Bean
   public String routingKey() {
-    return "r.#";
+    return "r.*." + stream();
   }
 
   @Bean
