+++ src/main/resources/application-localPublisher.properties
@@ -2,4 +2,5 @@ rabbitmq.host=localhost
 rabbitmq.port=5672
 rabbitmq.username=guest
 rabbitmq.password=guest
-restaurantId=8
+restaurantId=4
+server.port=8090
