+++ src/main/resources/application-localReceiver.properties
@@ -2,5 +2,5 @@ rabbitmq.host=localhost
 rabbitmq.port=5672
 rabbitmq.username=guest
 rabbitmq.password=guest
-restaurantId=3
+restaurantId=4
 spring.main.web-application-type=NONE
